requirements.txt
atomicwrites==1.3.0
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
future==0.17.1
hypothesis==3.88.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
joblib==0.13.1
MarkupSafe==1.1.0
more-itertools==6.0.0
numpy==1.16.1
onnx==1.3.0
pandas==0.24.1
pluggy==0.8.1
protobuf==3.6.1
py==1.7.0
pytest==4.3.0
pytext-nlp==0.1.4
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
scipy==1.2.0
six==1.12.0
tensorboardX==1.6
torch==1.0.1
torchtext==0.3.1
tqdm==4.30.0
typing==3.6.6
typing-extensions==3.7.2
urllib3==1.24.1
Werkzeug==0.14.1
